+++ tests/warp_drive/pycuda_tests/test_action_sampler.py
@@ -18,7 +18,6 @@ from warp_drive.utils.common import get_project_root
 from warp_drive.utils.constants import Constants
 from warp_drive.utils.data_feed import DataFeed
 
-pytorch_cuda_init_success = torch.cuda.FloatTensor(8)
 
 _CUBIN_FILEPATH = f"{get_project_root()}/warp_drive/cuda_bin"
 _ACTIONS = Constants.ACTIONS