+++ warp_drive/env_cpu_gpu_consistency_checker.py
@@ -23,7 +23,6 @@ from warp_drive.training.utils.data_loader import (
 from warp_drive.utils.constants import Constants
 from warp_drive.utils.data_feed import DataFeed
 
-pytorch_cuda_init_success = torch.cuda.FloatTensor(8)
 _OBSERVATIONS = Constants.OBSERVATIONS
 _ACTIONS = Constants.ACTIONS
 _REWARDS = Constants.REWARDS