+++ warp_drive/managers/pycuda_managers/pycuda_function_manager.py
@@ -13,7 +13,7 @@ import time
 from typing import Optional
 
 import numpy as np
-import pycuda.autoinit
+from warp_drive.utils import autoinit_pycuda
 import pycuda.driver as cuda_driver
 import torch
 from pycuda.compiler import SourceModule